Control which window is visible in full layout
Thanks to a bug in resize, when full was true, only the first window was
within the parent's geometry with the rest being offscreen. Now they all
the windows have the same dimension which brings us the question of
which should be on top. By default the windows are arranged in order
they were specified on the command line. If the "-m" was applied, this
order is revered. The ordering is also reversed if the new -l option is
used.